# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminToolCreateRequest": {
        "description": "`POST /admin/tools` —— 管理侧代创建，**允许指定 owner_id**（docs/03 §5.2）。",
        "properties": {
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "description_md": {
            "default": "",
            "maxLength": 100000,
            "title": "Description Md",
            "type": "string"
          },
          "name": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "owner_id": {
            "title": "Owner Id",
            "type": "integer"
          },
          "publish": {
            "default": false,
            "title": "Publish",
            "type": "boolean"
          },
          "summary": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "maxItems": 8,
            "title": "Tags",
            "type": "array"
          },
          "tool_type": {
            "$ref": "#/components/schemas/ToolType"
          },
          "visibility": {
            "$ref": "#/components/schemas/ToolVisibility",
            "default": "public"
          },
          "webapp_url": {
            "anyOf": [
              {
                "maxLength": 1024,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Webapp Url"
          }
        },
        "required": [
          "name",
          "summary",
          "tool_type",
          "owner_id"
        ],
        "title": "AdminToolCreateRequest",
        "type": "object"
      },
...
      "AdminToolListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminToolItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size",
          "pages"
        ],
        "title": "AdminToolListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ToolType": {
        "description": "docs/02 §3.8 `tools.tool_type`（D1：四种工具形态）",
        "enum": [
          "file",
          "webapp",
          "skill",
          "prompt"
        ],
        "title": "ToolType",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "selftool — 内网工具 / Skill 共享平台后端",
    "title": "selftool API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/recycle-bin": {
      "get": {
        "operationId": "list_recycle_bin_api_v1_admin_recycle_bin_get",
        "parameters": [
          {
            "description": "页码，从 1 开始",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "页码，从 1 开始",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "每页条数，1~200",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "每页条数，1~200",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminToolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "回收站列表",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/admin/tools": {
      "get": {
        "description": "FR-ADMIN-12：含所有状态、所有作者；支持状态/类型/分类/作者/可见性/时间范围筛选。",
        "operationId": "list_all_tools_api_v1_admin_tools_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "visibility",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Visibility"
            }
          },
          {
            "description": "按作者 username",
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按作者 username",
              "title": "Owner"
            }
          },
          {
            "in": "query",
            "name": "include_deleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Deleted",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "description": "页码，从 1 开始",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "页码，从 1 开始",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "每页条数，1~200",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "每页条数，1~200",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminToolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "全站工具列表",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "docs/03 §5.2：脚本批量导入用，**允许指定 owner_id**、可绕过 owner 校验。",
        "operationId": "create_tool_as_admin_api_v1_admin_tools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminToolCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminToolItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "管理侧代创建工具",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tools/{tool_id}/purge": {
      "delete": {
        "description": "**先删文件再删行** —— 顺序反了会留下无人引用的孤儿文件。",
        "operationId": "purge_tool_api_v1_admin_tools__tool_id__purge_delete",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeToolResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "彻底清除工具",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tools/{tool_id}/restore": {
      "post": {
        "operationId": "restore_tool_api_v1_admin_tools__tool_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminToolItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "从回收站还原",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tools/{tool_id}/transfer": {
      "post": {
        "description": "FR-IAM-11：重算双方配额 + 写 `transfer_owner` 审批留痕。",
        "operationId": "transfer_owner_api_v1_admin_tools__tool_id__transfer_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferOwnerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferOwnerResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "转移负责人",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/tools/{tool_id}/versions": {
      "post": {
        "description": "管理侧代上传的**说明性端点**。\n\n与 `/me/tools/{id}/versions` 共用 `version_service.create_version`，\n差别只在「谁有权调用」（这里由 `admin:all` 授权，不校验 owner）。\n\n由于 multipart 表单字段较多，实际实现复用同一套服务函数；\n这里只做参数透传，不重复业务逻辑。",
        "operationId": "upload_version_as_admin_api_v1_admin_tools__tool_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Upload Version As Admin Api V1 Admin Tools  Tool Id  Versions Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "管理侧代上传版本",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/me/tools": {
      "get": {
        "description": "含草稿/待审/驳回/下架，**排除软删除**。",
        "operationId": "list_my_tools_api_v1_me_tools_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          },
          {
            "description": "页码，从 1 开始",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "页码，从 1 开始",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "每页条数，1~200",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "每页条数，1~200",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyToolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "我的工具列表",
        "tags": [
          "me"
        ]
      },
      "post": {
        "description": "创建后是 `draft`，不进入审批队列；需要显式 submit。",
        "operationId": "create_tool_api_v1_me_tools_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "创建工具草稿",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/me/tools/{tool_id}/versions": {
      "get": {
        "operationId": "list_my_versions_api_v1_me_tools__tool_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionSummary"
                  },
                  "title": "Response List My Versions Api V1 Me Tools  Tool Id  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "我的工具版本列表（含待审与已驳回）",
        "tags": [
          "me"
        ]
      },
      "post": {
        "description": "见 `version_service.create_version` 的 12 步顺序说明。",
        "operationId": "upload_version_api_v1_me_tools__tool_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_version_api_v1_me_tools__tool_id__versions_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "上传新版本（multipart，12 步顺序）",
        "tags": [
          "me"
        ]
      }
    },
...
    "/api/v1/tools": {
      "get": {
        "description": "门户主列表。\n\n几个要点：\n\n- **`sort` 白名单校验**，非法值返回 400 `INVALID_SORT` 而不是 500；\n  排序表达式从 `SORT_EXPRESSIONS` 查表得到，**绝不把 `sort` 拼进 SQL**\n  （docs/03 §1.6）。\n- **`facets` 只在 `page == 1` 时返回**，翻页时整个键省略（docs/03 §3.3）。\n- 可见性判定下推到 SQL（EXISTS 子查询），不在 Python 里过滤\n  （docs/01 §4.3 性能要求）。\n- 未登录用户由 `portal.allow_anonymous_view` 决定 401 还是返回公开数据\n  （契约 §6），该判断在 `portal_access` 依赖里完成。",
        "operationId": "list_tools_api_v1_tools_get",
        "parameters": [
          {
            "description": "关键词，搜索名称、简介、详情、标签",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "关键词，搜索名称、简介、详情、标签",
              "title": "Q"
            }
          },
          {
            "description": "分类 slug，可重复传",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "分类 slug，可重复传",
              "title": "Category"
            }
          },
          {
            "description": "标签名，可重复传",
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "标签名，可重复传",
              "title": "Tag"
            }
          },
          {
            "description": "file/webapp/skill/prompt，可重复传",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "file/webapp/skill/prompt，可重复传",
              "title": "Type"
            }
          },
          {
            "description": "按作者 username 筛选",
            "in": "query",
            "name": "owner",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按作者 username 筛选",
              "title": "Owner"
            }
          },
          {
            "description": "hot（默认）/ new / name / -updated_at",
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "hot（默认）/ new / name / -updated_at",
              "title": "Sort"
            }
          },
          {
            "description": "页码，从 1 开始",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "页码，从 1 开始",
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "description": "每页条数，1~200",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "description": "每页条数，1~200",
              "maximum": 200,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "门户工具列表",
        "tags": [
          "portal"
        ]
      }
    },
...
    "/api/v1/tools/{slug}/versions": {
      "get": {
        "description": "对**有可见性权限的用户**开放（FR-VER-06）。\n\n`purged` 版本仍然出现在列表里但 `can_download=false` —— 界面显示「已归档」，\n而不是把它藏起来（那样用户会以为版本丢了）。",
        "operationId": "list_tool_versions_api_v1_tools__slug__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VersionSummary"
                  },
                  "title": "Response List Tool Versions Api V1 Tools  Slug  Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "版本列表（含待审、已驳回、已归档）",
        "tags": [
          "portal"
        ]
      }
    },
...
  }
}